[TPU 1VM] Consolidate all TPU ops related APIs into a single file
PiperOrigin-RevId: 338777305
