+++ tensorflow/core/tpu/kernels/tpu_mesh_state_interface.h
@@ -19,9 +19,8 @@ limitations under the License.
 
 #include "tensorflow/core/framework/resource_mgr.h"
 #include "tensorflow/core/protobuf/tpu/compile_metadata.pb.h"
-#include "tensorflow/core/tpu/kernels/tpu_compile_c_api.h"
-#include "tensorflow/core/tpu/kernels/tpu_mesh_state_c_api.h"
 #include "tensorflow/core/tpu/tpu_api.h"
+#include "tensorflow/core/tpu/tpu_ops_c_api.h"
 
 namespace tensorflow {
 
@@ -39,19 +38,19 @@ class TpuMeshStateInterface : public tensorflow::ResourceBase {
 
   ~TpuMeshStateInterface() override {
     if (mesh_state_ != nullptr) {
-      MeshStateApiFn()->TpuMeshState_FreeFn(mesh_state_);
+      OpsApiFn()->TpuMeshState_FreeFn(mesh_state_);
     }
   }
 
   static TpuMeshStateInterface* Create() {
-    return new TpuMeshStateInterface(MeshStateApiFn()->TpuMeshState_CreateFn());
+    return new TpuMeshStateInterface(OpsApiFn()->TpuMeshState_CreateFn());
   }
 
   const XLA_TpuMeshState* data() const { return mesh_state_; }
 
   tensorflow::TpuMeshCommonState* mesh_common_state() const {
     return static_cast<tensorflow::TpuMeshCommonState*>(
-        MeshStateApiFn()->TpuMeshState_MeshCommonStateFn(mesh_state_));
+        OpsApiFn()->TpuMeshState_MeshCommonStateFn(mesh_state_));
   }
 
   // Returns whether we should include the device assignment as a static field
@@ -63,8 +62,8 @@ class TpuMeshStateInterface : public tensorflow::ResourceBase {
     // Static device assignment enables XLA to perform certain optimization when
     // all cores are used in the replicated computation.
     return metadata.num_cores_per_replica() * metadata.num_replicas() ==
-           UtilApiFn()->TpuTopology_AvailableCoreCountFn(mesh_state_,
-                                                         tpu_core_type);
+           OpsApiFn()->TpuTopology_AvailableCoreCountFn(mesh_state_,
+                                                        tpu_core_type);
   }
 
   string DebugString() const override { return "TpuMeshStateInterface"; }
